pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "ocbrain"
version     = "2.0.0"
description = "OCBrain — local self-learning AI with modular expert models"
readme      = "README.md"
license     = { text = "MIT" }
requires-python = ">=3.11"

authors = [
    { name = "OCBrain", email = "hello@ocbrain.io" }
]

keywords = [
    "ai", "llm", "local", "self-learning", "ollama",
    "rag", "lora", "assistant", "ocbrain"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    "pydantic>=2.7.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
    "watchdog>=4.0.1",
    "spacy>=3.7.0,<4.0",
    "chromadb>=0.4.0,<1.0",
    "sentence-transformers>=3.0.0,<4.0",
    "trafilatura>=1.9.0",
    "feedparser>=6.0.11",
    "datasketch>=1.6.4",
    "pystray>=0.19.5",
    "Pillow>=10.3.0",
    "sqlalchemy>=2.0.30",
    "click>=8.1.7",
    "rich>=13.7.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
training = [
    "torch>=2.2.0",
    "trl>=0.9.0",
    "transformers>=4.40.0",
    "datasets>=2.17.0",
    "peft>=0.10.0",
    "bitsandbytes>=0.41.0",
]
voice = [
    "openai-whisper>=20231117",
    "pyttsx3>=2.90",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "keyboard>=0.13.5",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pyinstaller>=6.0.0",
]
all = [
    "ocbrain[training,voice]",
]

[project.scripts]
ocbrain       = "interface.cli:cli"
ocbrain-start = "main:main"

[project.urls]
Homepage      = "https://github.com/1h0lde4/OCBrain"
Repository    = "https://github.com/1h0lde4/OCBrain"
Issues        = "https://github.com/1h0lde4/OCBrain/issues"
Releases      = "https://github.com/1h0lde4/OCBrain/releases"
Documentation = "https://github.com/1h0lde4/OCBrain#readme"

[tool.setuptools.packages.find]
where   = ["."]
include = ["core*", "modules*", "learning*", "interface*", "install*"]

[tool.setuptools.package-data]
"*" = ["*.toml", "*.html", "*.css", "*.js", "*.png", "*.ico", "*.json"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]

[tool.setuptools.dynamic]
version = {file = "version.txt"}
